requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-jinja2==1.6
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kombu==5.4.2
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.0
textual==2.1.2
textual-dev==1.7.0
textual-serve==1.1.1
tornado==6.4.2
traitlets==5.14.3
tree-sitter==0.24.0
tree-sitter-bash==0.23.3
tree-sitter-css==0.23.2
tree-sitter-go==0.23.4
tree-sitter-html==0.23.2
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-json==0.24.8
tree-sitter-markdown==0.3.2
tree-sitter-python==0.23.6
tree-sitter-regex==0.24.3
tree-sitter-rust==0.23.2
tree-sitter-sql==0.3.7
tree-sitter-toml==0.7.0
tree-sitter-xml==0.7.0
tree-sitter-yaml==0.7.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.13
yarl==1.18.3
